meu_sistema_pagamento_banco/requirements.txt
flask
python-dotenv
mercadopago
psycopg2
